[CALCITE-3564] Unify function's operands type check in validation and runtime (DonnyZone)

+++ core/src/main/java/org/apache/calcite/sql/fun/SqlRandFunction.java
@@ -40,7 +40,7 @@ public class SqlRandFunction extends SqlFunction {
         SqlKind.OTHER_FUNCTION,
         ReturnTypes.DOUBLE,
         null,
-        OperandTypes.or(OperandTypes.NILADIC, OperandTypes.NUMERIC),
+        OperandTypes.or(OperandTypes.NILADIC, OperandTypes.INTEGER),
         SqlFunctionCategory.NUMERIC);
   }
